requirements.txt
# Lab Document Intelligence System
# HIPAA-Compliant Document Processing Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0  # For Azure App Service production deployment
python-multipart==0.0.6
jinja2==3.1.2

# Database
sqlalchemy>=2.0.36  # 2.0.36+ required for Python 3.13 support
pyodbc==5.0.1  # Required for Azure SQL Database
alembic>=1.13.0

# Azure Services
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
azure-monitor-opentelemetry==1.1.1

# AI/ML
openai==1.6.0  # For Azure OpenAI
pymupdf>=1.23.0  # PDF processing and conversion

# Security
cryptography==41.0.7
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
msal==1.28.0  # Microsoft Authentication Library for Entra ID OIDC
python3-saml>=1.16.0  # SAML 2.0 authentication for Entra ID SSO

# Validation (using versions with pre-built Windows wheels)
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0

# Fuzzy string matching
rapidfuzz>=3.0.0

# Monitoring
opencensus-ext-azure==1.1.13
applicationinsights==0.11.10

# Testing (dev dependencies)
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality (dev dependencies - optional, install separately if needed)
# pylint>=3.0.0
# flake8>=6.1.0
# black>=23.11.0
# bandit>=1.7.0
# safety>=3.0.0
